+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Bug fixes
 
 - Fixed bug in `Popover` where arrow could not be removed
+- The top checkbox in `StandardTable` deselects all if some are selected, as opposed to selecting all
 
 ## 8.1.0